moved adoption_model into src/ module update imports in test ignored some stuff
